+++ CHANGELOG.md
@@ -25,6 +25,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   the same name but with a `.NON_MATCHING` suffix is found on the same section.
   - The other suffixed symbol (`.NON_MATCHING`) is still retained in the section.
   - The suffixed symbol will have this member set to `false`.
+- Add support for the `decomp.yaml` specification from the
+  [`decomp_settings`](https://github.com/ethteck/decomp_settings) project on all
+  the provided CLI utilities.
+  - If a `decomp.yaml` file is detected, then every CLI argument that can be
+    inferred from that file will be be considered optional instead.
+  - Most CLI utilites will also add a new optional "version" argument to allow
+    picking the version to process from the `decomp.yaml` file. It defaults to
+    the first listed version.
 
 ### Changed